Restore files silently reverted by the historical WineCE merge
The original "Merge WineCE support into WineEX" merge (commit 023b5f82,
an unrelated-histories merge with a 2014-era Wine 1.6.2-based fork) had
no common ancestor with mainline, so git's merge resolution ended up
reverting large numbers of ordinary Wine source files to their ~2014
content wherever the WineCE side happened to touch them, well beyond
the 97 files that showed up as textual conflicts in the PR #12 rebase.

Identified every file that upstream wine-mirror did not touch between
the last sync point and PR #12's head, but that differs from that sync
point in our tree (i.e. changed only by our own fork history), then
excluded anything carrying an Xbox/UWP/WinRT/WineCE marker (genuine
WineEX platform-support work, left untouched) and any newly-added file
(also left untouched, since those have no upstream copy to conflict
with). The remaining ~2818 files matched a reversion signature (large,
mostly-deletion diffs versus current upstream, e.g. dlls/winex11.drv/
opengl.c had shrunk to reference an escape code struct that no longer
exists) and are restored to current wine-mirror content here.

Files with platform markers and all WineEX-original additions are
untouched by this commit.

+++ dlls/mountmgr.sys/cred.c
@@ -62,7 +62,7 @@ NTSTATUS query_symbol_file( void *args )
 
     if (!(query_cfstring = CFStringCreateWithFormat(kCFAllocatorDefault, NULL,
                                                     CFSTR("com_apple_xcode_dsym_uuids == \"%08X-%04X-%04X-%02X%02X-%02X%02X%02X%02X%02X%02X\""),
-                                                    (unsigned int)id->Data1, id->Data2, id->Data3, id->Data4[0],
+                                                    id->Data1, id->Data2, id->Data3, id->Data4[0],
                                                     id->Data4[1], id->Data4[2], id->Data4[3], id->Data4[4],
                                                     id->Data4[5], id->Data4[6], id->Data4[7] )))
         return STATUS_NO_MEMORY;
